[WIP] sof: crossover: Add Crossover component
This commit adds Crossover to the list of SOF components. A crossover
filter can be used to split an input to different frequency bands.
The number of outputs should be set statically in the topology. The user
then uses the control bytes to route the frequency bands to different
outputs. (similar to the demux component).

Signed-off-by: Sebastiano Carlucci <scarlucci@google.com>